changed dimensions calculations of players positions and tested on my phone moving on to button listeners

+++ FYP-FOOTBALLMANAGER/src/com/example/fyp_footballmanager/PlayerCoordinate.java
@@ -2,10 +2,10 @@ package com.example.fyp_footballmanager;
 
 public class PlayerCoordinate {
 
-	float xPos;
-	float yPos;
+	double xPos;
+	double yPos;
 	
-	public PlayerCoordinate(float x, float y) {
+	public PlayerCoordinate(double x, double y) {
 		xPos = x;
 		yPos = y;
 	}
